refactor: organize ProseMirror plugins
